fix(management): Better handling of settings
* Rename Portal and Console xxConfigEntity to xxxSettingsEntity
* Introduce a new ConsoleConfigEntity, which contains only the settings needed for console runtime and accessible when not authenticated
* deprecate former resources to get and save settings (/organization/orgID/console & organization/orgID/environments/envID/portal)
* add new resources to get and save settings  (/organization/orgID/settings & organization/orgID/environments/envID/settings)
Closes gravitee-io/issues#4804
