+++ gravitee-rest-api-management/gravitee-rest-api-management-rest/src/main/java/io/gravitee/rest/api/management/rest/resource/EnvironmentResource.java
@@ -296,4 +296,9 @@ public class EnvironmentResource extends AbstractResource {
     public NotifiersResource getNotifiersResource() {
         return resourceContext.getResource(NotifiersResource.class);
     }
+
+    @Path("settings")
+    public PortalSettingsResource getPortalSettingsResource() {
+        return resourceContext.getResource(PortalSettingsResource.class);
+    }
 }
