+++ gravitee-rest-api-model/src/main/java/io/gravitee/rest/api/model/settings/AbstractCommonSettingsEntity.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.rest.api.model.settings;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.gravitee.common.util.LinkedMultiValueMap;
+import io.gravitee.common.util.MultiValueMap;
+
+/**
+ * @author Florent CHAMFROY (florent.chamfroy at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+@JsonIgnoreProperties(value = {"baseURL"})
+public abstract class AbstractCommonSettingsEntity {
+
+    public static final String METADATA_READONLY = "readonly";
+    private Email email;
+
+    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
+    private MultiValueMap<String, String> metadata;
+
+    public AbstractCommonSettingsEntity() {
+        email = new Email();
+        metadata = new LinkedMultiValueMap<>();
+    }
+    // Getters & setters
+    public Email getEmail() {
+        return email;
+    }
+
+    public void setEmail(Email email) {
+        this.email = email;
+    }
+
+    public MultiValueMap<String, String> getMetadata() {
+        return metadata;
+    }
+
+    public void setMetadata(MultiValueMap<String, String> metadata) {
+        this.metadata = metadata;
+    }
+
+}
