+++ gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/java/io/gravitee/rest/api/portal/rest/resource/ConfigurationResource.java
@@ -72,7 +72,7 @@ public class ConfigurationResource extends AbstractResource {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public Response getPortalConfiguration() {
-        return Response.ok(configMapper.convert(configService.getPortalConfig())).build();
+        return Response.ok(configMapper.convert(configService.getPortalSettings())).build();
     }
 
     @GET
