+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/MediaServiceImpl.java
@@ -159,7 +159,7 @@ public class MediaServiceImpl implements MediaService {
 
     @Override
     public Long getMediaMaxSize() {
-        return Long.valueOf(configService.getPortalConfig().getPortal().getUploadMedia().getMaxSizeInOctet());
+        return Long.valueOf(configService.getPortalSettings().getPortal().getUploadMedia().getMaxSizeInOctet());
     }
 
     @Override
