+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/configuration/application/ApplicationTypeServiceImpl.java
@@ -17,9 +17,9 @@ package io.gravitee.rest.api.service.impl.configuration.application;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.gravitee.rest.api.model.parameters.PortalConfigEntity;
 import io.gravitee.rest.api.model.configuration.application.ApplicationTypeEntity;
 import io.gravitee.rest.api.model.configuration.application.ApplicationTypesEntity;
+import io.gravitee.rest.api.model.settings.PortalSettingsEntity;
 import io.gravitee.rest.api.service.ConfigService;
 import io.gravitee.rest.api.service.configuration.application.ApplicationTypeService;
 import io.gravitee.rest.api.service.exceptions.ApplicationTypeNotFoundException;
@@ -89,8 +89,8 @@ public class ApplicationTypeServiceImpl implements ApplicationTypeService {
     }
 
     public JsonNode getApplicationTypesConfiguration() {
-        PortalConfigEntity.Application applicationConfig = configService.getPortalConfig().getApplication();
-        PortalConfigEntity.Application.ApplicationTypes types = applicationConfig.getTypes();
+        PortalSettingsEntity.Application applicationConfig = configService.getPortalSettings().getApplication();
+        PortalSettingsEntity.Application.ApplicationTypes types = applicationConfig.getTypes();
         if (!applicationConfig.getRegistration().getEnabled()) {
             types.getBrowserType().setEnabled(false);
             types.getBackendToBackendType().setEnabled(false);
