+++ gravitee-rest-api-service/src/test/java/io/gravitee/rest/api/service/ConfigServiceTest.java
@@ -16,10 +16,11 @@
 package io.gravitee.rest.api.service;
 
 import io.gravitee.repository.management.model.Parameter;
-import io.gravitee.rest.api.model.parameters.ConsoleConfigEntity;
 import io.gravitee.rest.api.model.parameters.Key;
 import io.gravitee.rest.api.model.parameters.ParameterReferenceType;
-import io.gravitee.rest.api.model.parameters.PortalConfigEntity;
+import io.gravitee.rest.api.model.settings.ConsoleConfigEntity;
+import io.gravitee.rest.api.model.settings.ConsoleSettingsEntity;
+import io.gravitee.rest.api.model.settings.PortalSettingsEntity;
 import io.gravitee.rest.api.service.impl.ConfigServiceImpl;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -60,7 +61,7 @@ public class ConfigServiceTest {
     private NewsletterService newsletterService;
 
     @Test
-    public void shouldGetPortalConfig() {
+    public void shouldGetPortalSettings() {
 
         Map<String, List<String>> params = new HashMap<>();
         params.put(Key.PORTAL_AUTHENTICATION_FORCELOGIN_ENABLED.key(), singletonList("true"));
@@ -73,23 +74,23 @@ public class ConfigServiceTest {
         when(reCaptchaService.getSiteKey()).thenReturn("my-site-key");
         when(reCaptchaService.isEnabled()).thenReturn(true);
 
-        PortalConfigEntity portalConfig = configService.getPortalConfig();
+        PortalSettingsEntity portalSettings = configService.getPortalSettings();
 
-        assertNotNull(portalConfig);
-        assertEquals("force login", true, portalConfig.getAuthentication().getForceLogin().isEnabled());
-        assertEquals("scheduler notifications", Integer.valueOf(11), portalConfig.getScheduler().getNotificationsInSeconds());
-        assertEquals("analytics", Boolean.TRUE, portalConfig.getPortal().getAnalytics().isEnabled());
-        assertEquals("recaptcha siteKey", "my-site-key", portalConfig.getReCaptcha().getSiteKey());
-        assertEquals("recaptcha enabled", Boolean.TRUE, portalConfig.getReCaptcha().getEnabled());
-        assertEquals("plan security keyless", Boolean.TRUE, portalConfig.getPlan().getSecurity().getKeyless().isEnabled());
-        assertEquals("open api swagger enabled", Boolean.TRUE, portalConfig.getOpenAPIDocViewer().getOpenAPIDocType().getSwagger().isEnabled());
-        assertEquals("open api swagger default", "Swagger", portalConfig.getOpenAPIDocViewer().getOpenAPIDocType().getDefaultType());
-        assertEquals("api labels", 2, portalConfig.getApi().getLabelsDictionary().size());
-        assertEquals("cors exposed headers", 2, portalConfig.getCors().getExposedHeaders().size());
+        assertNotNull(portalSettings);
+        assertEquals("force login", true, portalSettings.getAuthentication().getForceLogin().isEnabled());
+        assertEquals("scheduler notifications", Integer.valueOf(11), portalSettings.getScheduler().getNotificationsInSeconds());
+        assertEquals("analytics", Boolean.TRUE, portalSettings.getPortal().getAnalytics().isEnabled());
+        assertEquals("recaptcha siteKey", "my-site-key", portalSettings.getReCaptcha().getSiteKey());
+        assertEquals("recaptcha enabled", Boolean.TRUE, portalSettings.getReCaptcha().getEnabled());
+        assertEquals("plan security keyless", Boolean.TRUE, portalSettings.getPlan().getSecurity().getKeyless().isEnabled());
+        assertEquals("open api swagger enabled", Boolean.TRUE, portalSettings.getOpenAPIDocViewer().getOpenAPIDocType().getSwagger().isEnabled());
+        assertEquals("open api swagger default", "Swagger", portalSettings.getOpenAPIDocViewer().getOpenAPIDocType().getDefaultType());
+        assertEquals("api labels", 2, portalSettings.getApi().getLabelsDictionary().size());
+        assertEquals("cors exposed headers", 2, portalSettings.getCors().getExposedHeaders().size());
     }
 
     @Test
-    public void shouldGetPortalConfigFromEnvVar() {
+    public void shouldGetPortalSettingsFromEnvVar() {
 
         Map<String, List<String>> params = new HashMap<>();
         params.put(Key.PORTAL_AUTHENTICATION_FORCELOGIN_ENABLED.key(), singletonList("true"));
@@ -107,16 +108,16 @@ public class ConfigServiceTest {
         when(environment.containsProperty(Key.PORTAL_ANALYTICS_ENABLED.key())).thenReturn(true);
         when(environment.containsProperty(Key.OPEN_API_DOC_TYPE_SWAGGER_ENABLED.key())).thenReturn(true);
 
-        PortalConfigEntity portalConfig = configService.getPortalConfig();
+        PortalSettingsEntity portalSettings = configService.getPortalSettings();
 
-        assertNotNull(portalConfig);
-        assertEquals("force login", true, portalConfig.getAuthentication().getForceLogin().isEnabled());
-        assertEquals("labels dictionary", 1, portalConfig.getApi().getLabelsDictionary().size());
-        assertEquals("scheduler notifications", Integer.valueOf(11), portalConfig.getScheduler().getNotificationsInSeconds());
-        assertEquals("analytics", Boolean.TRUE, portalConfig.getPortal().getAnalytics().isEnabled());
-        assertEquals("open api swagger enabled", Boolean.TRUE, portalConfig.getOpenAPIDocViewer().getOpenAPIDocType().getSwagger().isEnabled());
-        assertEquals("cors exposed headers", 1, portalConfig.getCors().getExposedHeaders().size());
-        List<String> readonlyMetadata = portalConfig.getMetadata().get(PortalConfigEntity.METADATA_READONLY);
+        assertNotNull(portalSettings);
+        assertEquals("force login", true, portalSettings.getAuthentication().getForceLogin().isEnabled());
+        assertEquals("labels dictionary", 1, portalSettings.getApi().getLabelsDictionary().size());
+        assertEquals("scheduler notifications", Integer.valueOf(11), portalSettings.getScheduler().getNotificationsInSeconds());
+        assertEquals("analytics", Boolean.TRUE, portalSettings.getPortal().getAnalytics().isEnabled());
+        assertEquals("open api swagger enabled", Boolean.TRUE, portalSettings.getOpenAPIDocViewer().getOpenAPIDocType().getSwagger().isEnabled());
+        assertEquals("cors exposed headers", 1, portalSettings.getCors().getExposedHeaders().size());
+        List<String> readonlyMetadata = portalSettings.getMetadata().get(PortalSettingsEntity.METADATA_READONLY);
         assertEquals("Config metadata size", 5, readonlyMetadata.size());
         assertTrue("Config metadata contains AUTHENTICATION_FORCELOGIN_ENABLED", readonlyMetadata.contains(Key.PORTAL_AUTHENTICATION_FORCELOGIN_ENABLED.key()));
         assertTrue("Config metadata contains API_LABELS_DICTIONARY", readonlyMetadata.contains(Key.API_LABELS_DICTIONARY.key()));
@@ -126,14 +127,36 @@ public class ConfigServiceTest {
     }
 
     @Test
-    public void shouldCreatePortalConfig() {
-        PortalConfigEntity portalConfigEntity = new PortalConfigEntity();
-        portalConfigEntity.getPortal().setUrl("ACME");
+    public void shouldCreatePortalSettings() {
+        PortalSettingsEntity portalSettingsEntity = new PortalSettingsEntity();
+        portalSettingsEntity.getPortal().setUrl("ACME");
         when(mockParameterService.save(PORTAL_URL, "ACME", "DEFAULT", ParameterReferenceType.ENVIRONMENT)).thenReturn(new Parameter());
 
-        configService.save(portalConfigEntity);
+        configService.save(portalSettingsEntity);
 
         verify(mockParameterService, times(1)).save(PORTAL_URL, "ACME", "DEFAULT", ParameterReferenceType.ENVIRONMENT);
+    }
+
+    @Test
+    public void shouldGetConsoleSettings() {
+
+        Map<String, List<String>> params = new HashMap<>();
+        params.put(COMPANY_NAME.key(), singletonList("ACME"));
+        params.put(Key.CONSOLE_SCHEDULER_NOTIFICATIONS.key(), singletonList("11"));
+        params.put(Key.ALERT_ENABLED.key(), singletonList("true"));
+
+        when(mockParameterService.findAll(any(List.class), eq("DEFAULT"), eq(ParameterReferenceType.ORGANIZATION))).thenReturn(params);
+        when(reCaptchaService.getSiteKey()).thenReturn("my-site-key");
+        when(reCaptchaService.isEnabled()).thenReturn(true);
+
+        ConsoleSettingsEntity consoleSettings = configService.getConsoleSettings();
+
+        assertNotNull(consoleSettings);
+        assertEquals("scheduler notifications", Integer.valueOf(11), consoleSettings.getScheduler().getNotificationsInSeconds());
+        assertEquals("recaptcha siteKey", "my-site-key", consoleSettings.getReCaptcha().getSiteKey());
+        assertEquals("alerting enabled", Boolean.TRUE, consoleSettings.getAlert().getEnabled());
+        assertEquals("recaptcha enabled", Boolean.TRUE, consoleSettings.getReCaptcha().getEnabled());
+        assertEquals("cors exposed headers", 2, consoleSettings.getCors().getExposedHeaders().size());
     }
 
     @Test
@@ -155,11 +178,10 @@ public class ConfigServiceTest {
         assertEquals("recaptcha siteKey", "my-site-key", consoleConfig.getReCaptcha().getSiteKey());
         assertEquals("alerting enabled", Boolean.TRUE, consoleConfig.getAlert().getEnabled());
         assertEquals("recaptcha enabled", Boolean.TRUE, consoleConfig.getReCaptcha().getEnabled());
-        assertEquals("cors exposed headers", 2, consoleConfig.getCors().getExposedHeaders().size());
     }
 
     @Test
-    public void shouldGetConsoleConfigFromEnvVar() {
+    public void shouldGetConsoleSettingsFromEnvVar() {
 
         Map<String, List<String>> params = new HashMap<>();
         params.put(COMPANY_NAME.key(), singletonList("ACME"));
@@ -174,24 +196,24 @@ public class ConfigServiceTest {
         when(environment.containsProperty(eq(Key.CONSOLE_AUTHENTICATION_LOCALLOGIN_ENABLED.key()))).thenReturn(true);
         when(environment.containsProperty(Key.CONSOLE_SCHEDULER_NOTIFICATIONS.key())).thenReturn(true);
 
-        ConsoleConfigEntity consoleConfig = configService.getConsoleConfig();
+        ConsoleSettingsEntity consoleSettings = configService.getConsoleSettings();
 
-        assertNotNull(consoleConfig);
-        assertEquals("scheduler notifications", Integer.valueOf(11), consoleConfig.getScheduler().getNotificationsInSeconds());
-        assertEquals("cors exposed headers", 1, consoleConfig.getCors().getExposedHeaders().size());
-        List<String> readonlyMetadata = consoleConfig.getMetadata().get(PortalConfigEntity.METADATA_READONLY);
+        assertNotNull(consoleSettings);
+        assertEquals("scheduler notifications", Integer.valueOf(11), consoleSettings.getScheduler().getNotificationsInSeconds());
+        assertEquals("cors exposed headers", 1, consoleSettings.getCors().getExposedHeaders().size());
+        List<String> readonlyMetadata = consoleSettings.getMetadata().get(PortalSettingsEntity.METADATA_READONLY);
         assertEquals("Config metadata size", 2, readonlyMetadata.size());
         assertTrue("Config metadata contains CONSOLE_AUTHENTICATION_LOCALLOGIN_ENABLED", readonlyMetadata.contains(Key.CONSOLE_AUTHENTICATION_LOCALLOGIN_ENABLED.key()));
         assertTrue("Config metadata contains CONSOLE_SCHEDULER_NOTIFICATIONS", readonlyMetadata.contains(Key.CONSOLE_SCHEDULER_NOTIFICATIONS.key()));
     }
 
     @Test
-    public void shouldCreateConsoleConfig() {
-        ConsoleConfigEntity consoleConfigEntity = new ConsoleConfigEntity();
-        consoleConfigEntity.getAlert().setEnabled(true);
+    public void shouldCreateConsoleSettings() {
+        ConsoleSettingsEntity consoleSettingsEntity = new ConsoleSettingsEntity();
+        consoleSettingsEntity.getAlert().setEnabled(true);
         when(mockParameterService.save(ALERT_ENABLED, "true", "DEFAULT", ParameterReferenceType.ORGANIZATION)).thenReturn(new Parameter());
 
-        configService.save(consoleConfigEntity);
+        configService.save(consoleSettingsEntity);
 
         verify(mockParameterService, times(1)).save(ALERT_ENABLED, "true", "DEFAULT", ParameterReferenceType.ORGANIZATION);
     }
